\name{colorado} 

\alias{colorado}

\title{
	Colorado precipitation data.
} 

\description{
	Data set related to Colorado precipitation in May 1997. Taken from http://www.image.ucar.edu/Data/US.monthly.met/.  Provides a list containing a transformed precipitation variance contained in the vector \code{odata}, as well as the associated longitude/latitude are in \code{ocoords}.
} 

\usage{
	data(colorado)
} 

\author{Joshua French} 

\examples{
	data(colorado)
}