\name{ExceedanceTools-package}
\alias{ExceedanceTools-package}
\alias{ExceedanceTools}
\docType{package}
\title{
Confidence regions for exceedance regions.
}
\description{
A package for constructing confidence regions for exceedance regions.
}
\author{
Joshua P. French
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\keyword{ exceedance }
\keyword{ hotspot }

\seealso{
SpatialTools
}
\examples{
rnorm(10)
}
