% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExceedanceTools.R
\docType{package}
\name{ExceedanceTools}
\alias{ExceedanceTools}
\alias{ExceedanceTools-package}
\title{ExceedanceTools.}
\description{
A package to create confidence or credible regions for
the exceedance regions/excursion sets of spatial data.
}
