% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{sdata}
\alias{sdata}
\title{Synthetic data}
\format{A data frame.}
\usage{
data(sdata)
}
\description{
A synthetic data set for use in examples. A 100x3 data frame with vectors \code{x1} and \code{x2} (specifying spatial location) and \code{y}, the response.
}
\author{
Joshua French
}

