% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorado-data.R
\docType{data}
\name{colorado}
\alias{colorado}
\title{Colorado precipitation data}
\format{
A list.
}
\source{
National Center for Atmospheric Research
}
\usage{
data(colorado)
}
\description{
Data related to Colorado precipitation in
  May 1997. Taken from
  \url{https://www.image.ucar.edu/Data/US.monthly.met/}.
  Data is contained in a list with components
  \code{odata} (containing a transformed precipitation
  variable) and \code{ocoords} containing the longitude
  and latitude of the associated sites.
}
\author{
Joshua French
}
