\name{ExceedanceTools-package}
\alias{ExceedanceTools-package}
\alias{ExceedanceTools}
\docType{package}
\title{
Confidence regions for exceedance regions.
}
\description{
A package for constructing confidence regions for exceedance regions.
}
\author{
Joshua P. French
}
\references{
French, J.P. and Sain, S.R. Spatio-Temporal Exceedance Locations and Confidence Regions. Annals of Applied Statistics.  Prepress.
}

\keyword{ package }
\keyword{ exceedance }
\keyword{ hotspot }

\seealso{
SpatialTools
}
\examples{
rnorm(10)
}
