\name{sdata} 

\alias{sdata}
\alias{y}
\alias{coords}

\title{
	Synthetic data set.
} 

\description{
	A synthetic data sets for use in examples.  Contains \code{coords}, a 100 x 2 matrix of coordinates for the responses, and \code{y}, a vector of length 100 of observed responses.
} 

\usage{
	data(sdata)
} 

\author{Joshua French} 

\examples{
	data(sdata)
}