% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdata.R
\docType{data}
\name{sdata}
\alias{sdata}
\title{Synthetic data}
\format{
A data frame.
}
\usage{
data(sdata)
}
\description{
A synthetic data set for use in examples. A 100x3 data frame with vectors \code{x1} and \code{x2} (specifying spatial location) and \code{y}, the response.
}
\author{
Joshua French
}
